refactor : org 패키지명 -> calculator 패키지 명으로 변경 및 자바 버전 11 -> 15로 변경
