+++ build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'java'
 }
 
-group 'org'
+group 'calculator'
 version '1.0-SNAPSHOT'
 
 repositories {
