+++ src/main/java/calculator/domain/repository/CalculationRepository.java
@@ -1,6 +1,6 @@
-package org.domain.repository;
+package calculator.domain.repository;
 
-import org.domain.model.Calculator;
+import calculator.domain.model.Calculator;
 
 import java.util.*;
