+++ src/main/java/calculator/error/ResponseErrorFormat.java
@@ -1,4 +1,4 @@
-package org.error;
+package calculator.error;
 
 public enum ResponseErrorFormat {
